requirements.txt
streamlit>=1.32.0
requests>=2.31.0
Pillow>=10.0.0
python-docx>=0.8.11
reportlab>=4.0.0